refactor: minimal api construct
- tooling should handle special interactions with constructs
- the constructs themselves should be used primarily to store info
- this reduces possible issues with JIT compiling
