+++ packages/ant-stack/src/utils/index.ts
@@ -1,2 +1,9 @@
-export * from "./directories";
-export * from "./wait-for-stack-idle";
+export * from "./deep-partial.js";
+export * from "./directories.js";
+export * from "./execute-jit.js";
+export * from "./express-apigateway.js";
+export * from "./files.js";
+export * from "./index.js";
+export * from "./package-manager.js";
+export * from "./static-analysis.js";
+export * from "./wait-for-stack-idle.js";
